Fix integer dtype for index array in AETC legacy code
Use dtype=int for zeros(1) to prevent float promotion when hstacking
with integer covariate_subset, which caused np.ix_ to fail with
IndexError: arrays used as indices must be of integer (or boolean) type

+++ pyapprox/multifidelity/etc.py
@@ -82,7 +82,7 @@ class AETC:
         cov, means = oracle_stats[:2]
         Sigma_S = cov[np.ix_(covariate_subset + 1, covariate_subset + 1)]
         Sp_subset = self._bkd.hstack(
-            [self._bkd.zeros(1), covariate_subset + 1]
+            [self._bkd.zeros(1, dtype=int), covariate_subset + 1]
         )
         x_Sp = self._bkd.vstack(
             (self._bkd.ones(1), means[covariate_subset + 1])
